requirements.txt
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.0.8
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.5
Cython==3.0.10
debugpy==1.6.3
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
onnx==1.16.1
openai==1.37.1
opencv-python-headless==4.10.0.84
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
playsound==1.3.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32==304
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.5.1
starlette==0.37.2
tornado==6.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
wcwidth==0.2.5
websockets==12.0
